showProductPriceStateで金額を小数点まで表示するように対応した

+++ vendingmachine/src/vendingmachine/VendingMachine.java
@@ -54,7 +54,7 @@ public class VendingMachine {
 			return;
 		}
 		this.selectedProduct = product;
-		ShowProductPriceState.getInstance().productPrice(product.price().intValue());;
+		ShowProductPriceState.getInstance().productPrice(product.price());;
 		displayPanel.pushProductButton();
 	}
